Change naming around Algolia index fake
The double that replaces real calls to Algolia in our test suite can be
referred to by a slightly more precise name - a fake.

Rename accordingly, and also change references to 'algolia directory'
to 'algolia index' for consistency.
