+++ spec/features/self_service/offices_index_spec.rb
@@ -1,5 +1,5 @@
 RSpec.feature 'The self service firm offices list page', :inline_job_queue do
-  include_context 'algolia directory double'
+  include_context 'algolia index fake'
 
   let(:offices_index_page) { SelfService::OfficesIndexPage.new }
   let(:sign_in_page) { SignInPage.new }
